% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_functions.r
\name{contrast_each_group_to_the_rest}
\alias{contrast_each_group_to_the_rest}
\title{contrast_each_group_to_the_rest}
\usage{
contrast_each_group_to_the_rest(dataset_se, dataset_name, groups2test = NA)
}
\arguments{
\item{dataset_se}{Summarised experiment object containing count data. Also
requires 'ID' and 'group' to be set within the cell information
(see \code{colData()})}

\item{dataset_name}{Short, meaningful name for this dataset/experiment.}

\item{groups2test}{An optional character vector specificing specific groups to check.
By default (set to NA), all groups will be tested.}
}
\value{
A tibble, the de_table (differential expression table).  This table
is a core interal summary of each individual dataset, which is
used for the cross-dataset comparisons. So save this output to pass to other
functions.
}
\description{
\code{contrast_each_group_to_the_rest} produces a table of differential
expression results, where each group (cluster) is compared to the rest of the
cells.
}
\details{
Note that this function is \emph{slow}, because it runs the differential
expression. It only needs to be run once per dataset (unless group labels
change)
}
\examples{
de_table.demo_query  <- contrast_each_group_to_the_rest(demo_query_se, "a_demo_query")

}
