% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{de_table.demo_ref}
\alias{de_table.demo_ref}
\title{Demo ref de table}
\format{An object of class \code{data.frame} with 800 rows and 13 columns.}
\usage{
de_table.demo_ref
}
\description{
Small example dataset that is the output of
\link[celaref]{contrast_each_group_to_the_rest}. It contains the results
of each group compared to the rest of the sample (ie within sample 
differential expression)
}
\keyword{datasets}
