% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo_cell_info_table}
\alias{demo_cell_info_table}
\title{Demo cell info table}
\format{An object of class \code{data.frame} with 515 rows and 4 columns.}
\usage{
demo_cell_info_table
}
\value{
An example cell info table
}
\description{
Sample sheet table listing each cell, its assignd cluster/group, and 
any other information that might be interesting (replicate, individual e.t.c)
}
\keyword{datasets}
