% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo_counts_matrix}
\alias{demo_counts_matrix}
\title{Demo count matrix}
\format{An object of class \code{matrix} with 200 rows and 514 columns.}
\usage{
demo_counts_matrix
}
\description{
Counts matrix for a small, demo example datasets. Raw counts of 
reads per gene (row) per cell (column).
}
\keyword{datasets}
