% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo_gene_info_table}
\alias{demo_gene_info_table}
\title{Demo gene info table}
\format{An object of class \code{data.frame} with 200 rows and 2 columns.}
\usage{
demo_gene_info_table
}
\value{
An example table of genes.
}
\description{
Extra table of gene-level information for the demo example dataset.
Can contain anything as long as theres a unique gene id.
}
\keyword{datasets}
