% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo_microarray_expr}
\alias{demo_microarray_expr}
\title{Demo microarray expression table}
\format{An object of class \code{matrix} with 200 rows and 20 columns.}
\usage{
demo_microarray_expr
}
\description{
Microarray-style expression table for the demo example dataset. 
Rows are genes, columns are samples, as per counts matrix.
}
\keyword{datasets}
