% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo_microarray_sample_sheet}
\alias{demo_microarray_sample_sheet}
\title{Demo microarray sample sheet table}
\format{An object of class \code{data.frame} with 20 rows and 2 columns.}
\usage{
demo_microarray_sample_sheet
}
\description{
Microarray sample sheet table for the demo example dataset. 
Contains array identifiers, their group and any other information that could
be useful.
}
\keyword{datasets}
