% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo_query_se}
\alias{demo_query_se}
\title{Demo query se (summarizedExperiment)}
\format{An object of class \code{SummarizedExperiment} with 100 rows and 106 columns.}
\usage{
demo_query_se
}
\description{
A summarisedExperiment object loaded from demo info tables, for a query set.
}
\keyword{datasets}
