% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{demo_ref_se}
\alias{demo_ref_se}
\title{Demo reference se (summarizedExperiment)}
\format{An object of class \code{SummarizedExperiment} with 100 rows and 94 columns.}
\usage{
demo_ref_se
}
\description{
A summarisedExperiment object loaded from demo info tables, for a reference
set.
}
\keyword{datasets}
