% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{find_within_match_differences}
\alias{find_within_match_differences}
\title{find_within_match_differences}
\usage{
find_within_match_differences(de_table.ref.marked, matches, the_test_group,
  the_test_dataset, the_ref_dataset, the_pval)
}
\arguments{
\item{de_table.ref.marked}{see make_ref_similarity_names_for_group}

\item{matches}{see make_ref_similarity_names_for_group}

\item{the_test_group}{see make_ref_similarity_names_for_group}

\item{the_test_dataset}{see make_ref_similarity_names_for_group}

\item{the_ref_dataset}{see make_ref_similarity_names_for_group}

\item{the_pval}{see make_ref_similarity_names_for_group}
}
\value{
String of within match differences
}
\description{
Internal function to find if there are significant difference between the
distribitions, when there are multiple match groups.
}
\details{
For use by make_ref_similarity_names_for_group
}
\seealso{
\code{\link{make_ref_similarity_names_for_group}}
}
