% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_functions.r
\name{get_counts_index}
\alias{get_counts_index}
\title{get_counts_index}
\usage{
get_counts_index(n_assays, assay_names)
}
\arguments{
\item{n_assays}{How many assays are there? ie: length(assays(dataset_se))}

\item{assay_names}{What are the assays called? ie: names(assays(dataset_se))}
}
\value{
The index of an assay in assays called 'counts', or, if there's just
     the one unnamed assay - happily assume that that is counts.
}
\description{
\code{get_counts_index} is an internal utility function to find out where 
the counts are (if anywhere.). Stops if there's no assay called 'counts',
(unless there is only a single unnamed assay).
}
