% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_functions.r
\name{get_inner_or_outer_ci}
\alias{get_inner_or_outer_ci}
\title{get_inner_or_outer_ci}
\usage{
get_inner_or_outer_ci(fc, ci.hi, ci.lo, get_inner = TRUE)
}
\arguments{
\item{fc}{Fold-change}

\item{ci.hi}{Higher fold-change CI (numerically)}

\item{ci.lo}{smaller fold-change CI (numerically)}

\item{get_inner}{If TRUE, get the more conservative inner CI, else the bigger
outside one.}
}
\value{
inner or outer CI from \bold{ci.hi} or \bold{ci.low}
}
\description{
Given a fold-change, and high and low confidence interval (where lower <
higher), pick the innermost/most conservative one.
}
