% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{get_matched_stepped_mwtest_res_table}
\alias{get_matched_stepped_mwtest_res_table}
\title{get_matched_stepped_mwtest_res_table}
\usage{
get_matched_stepped_mwtest_res_table(mwtest_res_table.this, the_pval)
}
\arguments{
\item{mwtest_res_table.this}{Combined output of 
\code{\link{get_ranking_and_test_results}}}

\item{the_pval}{Pvalue threshold}
}
\value{
Stepped pvalues string
}
\description{
Internal function to grab a table of the matched group(s).
}
\details{
For use by make_ref_similarity_names_for_group
}
\seealso{
\code{\link{make_ref_similarity_names_for_group}}
}
