% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{get_reciprocal_matches}
\alias{get_reciprocal_matches}
\title{get_reciprocal_matches}
\usage{
get_reciprocal_matches(mwtest_res_table.recip, de_table.recip.marked,
  the_pval)
}
\arguments{
\item{mwtest_res_table.recip}{Combined output of 
\code{\link{get_ranking_and_test_results}} for reciprocal test - 
ref vs query.}

\item{de_table.recip.marked}{Recriprocal ref vs query de_table.ref.marked}

\item{the_pval}{See make_ref_similarity_names_using_marked}
}
\value{
List of table of reciprocal matches tested from reference to query.
}
\description{
Internal function to run a bionomial test of 
median test rank > 0.5 (random).
}
\details{
For use by make_ref_similarity_names_using_marked
}
\seealso{
\code{\link{make_ref_similarity_names_using_marked}}
}
