% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasting_functions.r
\name{get_the_up_genes_for_all_possible_groups}
\alias{get_the_up_genes_for_all_possible_groups}
\title{get_the_up_genes_for_all_possible_groups}
\usage{
get_the_up_genes_for_all_possible_groups(de_table.test, de_table.ref,
  test_dataset_name, rankmetric = "TOP100_LOWER_CI_GTE1")
}
\arguments{
\item{de_table.test}{A differential expression table of the query experiment,
as generated from \code{\link[celaref]{contrast_each_group_to_the_rest}}}

\item{de_table.ref}{A differential expression table of the reference dataset,
as generated from \code{\link[celaref]{contrast_each_group_to_the_rest}}}

\item{test_dataset_name}{A short meaningful name for the query experiment.}

\item{rankmetric}{Placeholder for support of different ranking methods, but only the default supported. Omit.}
}
\value{
\emph{de_table.marked} This will alsmost be a subset of \bold{de_table.ref}, 
with an added column \emph{test_group} set to the query groups, and 
\emph{test_dataset} set to \bold{test_dataset_name}.

If nothing passes the rankmetric criteria, a warning is thrown and NA is 
returned. (This can be a genuine inability to pick out the 
representative 'up' genes, or due to some problem in the analysis)
}
\description{
For the most overrepresented genes of each group in the test 
dataset, get their rankings in all the groups of the reference dataset.
}
\details{
This is effectively a subset of the reference data, 'marked' with the 'top'
genes that represent the groups in the query data. The 
distribution of the \emph{rescaled ranks} of these marked genes in each 
reference data group indicate how similar they are to the query group.

This function is simply a conveinent wrapper for 
\code{\link[celaref]{get_the_up_genes_for_group}} that merges output for 
each group in the query into one table.
}
\examples{
de_table.marked.query_vs_ref <- get_the_up_genes_for_all_possible_groups(
   test_dataset_name="query",
   de_table.test=de_table.demo_query ,
   de_table.ref=de_table.demo_ref )

}
\seealso{
\code{\link[celaref]{get_the_up_genes_for_group}} Function for testing a single group.
}
