% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{get_vs_random_pval}
\alias{get_vs_random_pval}
\title{get_vs_random_pval}
\usage{
get_vs_random_pval(de_table.ref.marked, the_group, the_test_group)
}
\arguments{
\item{de_table.ref.marked}{see make_ref_similarity_names_for_group}

\item{the_group}{Reference group name}

\item{the_test_group}{Test group name
#'}
}
\value{
Pvalue result of a binomial test of each 'top gene' being greater than
the theoretical random median rank of 0.5 (halfway).
}
\description{
Internal function to run a bionomial test of median test rank > 0.5 (random).
}
\details{
For use by make_ref_similarity_names_for_group
}
\seealso{
\code{\link[celaref]{make_ref_similarity_names_for_group}}
}
