% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_helper_functions.r
\name{load_se_from_files}
\alias{load_se_from_files}
\title{load_se_from_files}
\usage{
load_se_from_files(counts_file, cell_info_file, gene_info_file = NA,
  group_col_name = "group")
}
\arguments{
\item{counts_file}{A tab-separated file of a matrix of read counts for each gene
(row) and each cell (column). First column should be gene names,
and top row cell ids.}

\item{cell_info_file}{Tab-separated text file of cell information.
Columns must have names. If there is a column labelled
\emph{cell_sample}, that will be used as the unique cell identifiers. If not,
the first column is assumed to be cell identifiers, and will be copied to a
new feild labelled \emph{cell_sample}.
Similarly - the clusters of these cells should be listed in one column -
which can be called 'group' (case-sensitive) or specified with
\bold{group_col_name}. \emph{Minimal file format: <cell_sample> <group>.}}

\item{gene_info_file}{Optional tab-separated text file of gene information.
Columns must have names. If there is a column labelled
\emph{ID}, that will be used as the gene identifiers (they must be unique!).
If not, the first column is assumed to be a gene identifier, and will be copied to a
new feild labelled \emph{ID}. Must match all rownames in counts_matrix file.
If omitted, ID wll be generated from counts file rownames. Default=NA}

\item{group_col_name}{Name of the column in \bold{cell_info_file} containing
the cluster/group that each cell belongs to. Case-sensitive. Default='group'}
}
\value{
A SummarisedExperiment object containing the count data, cell info
and gene info.
}
\description{
Create a SummarizedExperiment object (dataset_se) from files.
}
\details{
This convenience function makes a SummarizedExperiment object in a form that
should work for celaref functions. Specifically, that means it will have an
'ID' feild for genes (view with \code{rowData(dataset_se)}), and both
'cell_sample' and 'group' feild for cells (view with
\code{colData(dataset_se)}). See parameters for detail.
Additionally, the counts will be an integer matrix (not a
sparse matrix), and the \emph{group} feild (but not \emph{cell_sample}
or \emph{ID}) will be a factor.

Note that data will be subsetted to cells present in both the counts matrix
and cell info files, this is handy for loading subsets of cells.
However, if \bold{gene_info_file} is defined, all genes must match exactly.
}
\examples{

# Use demo data files
counts_filepath    <- system.file("extdata", "sim_query_counts.tab",    package = "celaref")
cell_info_filepath <- system.file("extdata", "sim_query_cell_info.tab", package = "celaref")
gene_info_filepath <- system.file("extdata", "sim_query_gene_info.tab", package = "celaref")

demo_se <- load_se_from_files(counts_file=counts_filepath, cell_info_file=cell_info_filepath)

# Same counts, but with some extra info on the genes stored
demo_se <- load_se_from_files(counts_file=counts_filepath, cell_info_file=cell_info_filepath, gene_info_filepath=gene_info_filepath )

}
\seealso{
\code{\link[RangedSummarizedExperiment-class]{SummarizedExperiment}} For general doco on the SummarizedExperiment objects.

For reading in these objects in a form that works with this function see \code{\link[celaref]{lload_se_from_files}} or \code{\link[celaref]{load_dataset_10Xdata}}
}
