% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{make_ref_similarity_names_for_groups}
\alias{make_ref_similarity_names_for_groups}
\title{make_ref_similarity_names_for_groups}
\usage{
make_ref_similarity_names_for_groups(de_table.ref.marked, the_test_dataset,
  the_ref_dataset, pval = 0.01)
}
\arguments{
\item{de_table.ref.marked}{The output of \code{\link{get_the_up_genes_for_all_possible_groups}} for the contrast of interest.}

\item{the_test_dataset}{A short meaningful name for the experiment. (Should match \emph{test_dataset} column in \bold{de_table.marked})}

\item{the_ref_dataset}{A short meaningful name for the experiment. (Should match \emph{dataset} column in \bold{de_table.marked})}

\item{pval}{Differences between the rescaled ranking distribution of 'top'
genes on different reference groups are tested with a Mann-Whitney U test. If 
they are \emph{significantly different}, only the top group(s) are reported. 
ie. A more stringent \bold{pval} is more likely to report multiple 
reference groups. 
This parameter unlikely to need change. Default = 0.01.}
}
\value{
A table of automagically-generated labels for each query group, given 
their similarity to reference groups. Short_lab and long_lab are just 
different versions of the name.
}
\description{
Given the comparison between the two dataset, construct some sensible labels 
for the groups in the query group.
}
\details{
This function aims to report a) the top most similar reference group, if 
there's a clear frontrunner, b) A list of multiple similar groups if they 
have similar similarity, or c) 'No similarity', if there is none.

Each group is named according to the following rules. Testing for significant 
(smaller) differences with a one-directional Mann-Whitney U test on their rescaled ranks:
\enumerate{
  \item The first (as ranked by median rescaled rank) reference group is 
  significantly more similar than the next: Report \emph{first only}.
  \item When comparing differences betwen groups stepwise ranked by 
  median rescaled rank - no group is significantly different to its 
  neighbour: Report \emph{no similarity}
  \item There's no significant differences in the stepwise comparisons 
  of the first N reference groups - but there is a significant 
  difference later on : Report \emph{multiple group similarity}
}

The similarity is formatted into a group label. Where there are 
multiple similar groups, they're listed from most to least similar by their 
median ranks.
 
For instance, a query dataset of clusters c1, c2 and c3 againsts a cell-type
labelled reference datatset might get names like:
E.g.
\itemize{
  \item c1:macrophage
  \item c2:endotheial|mesodermal
  \item c3:no_similarity
}
}
\examples{

de_table.demo_query <- contrast_each_group_to_the_rest(demo_query_se, "demo_query", num_cores=2)
de_table.demo_ref   <- contrast_each_group_to_the_rest(demo_ref_se,   "demo_ref", num_cores=2)
de_table.marked.query_vs_ref <- get_the_up_genes_for_all_possible_groups(
     de_table.demo_query, de_table.demo_ref, 'demo_query')


}
\seealso{
\code{\link[celaref]{get_the_up_genes_for_all_possible_groups}} To prepare the \bold{de_table.ref.marked} input.
}
