% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{make_ref_similarity_names_for_groups_ks}
\alias{make_ref_similarity_names_for_groups_ks}
\title{make_ref_similarity_names_for_groups_ks}
\usage{
make_ref_similarity_names_for_groups_ks(de_table.ref.marked, the_test_dataset,
  the_ref_dataset, ks_pval = 0.01, median_rank_threshold = 0.25)
}
\arguments{
\item{the_test_dataset}{A short meaningful name for the experiment. (Should match \emph{test_dataset} column in \bold{de_table.marked})}

\item{the_ref_dataset}{A short meaningful name for the experiment. (Should match \emph{dataset} column in \bold{de_table.marked})}

\item{ks_pval}{When multiple reference groups pass the median rank threshold, 
the difference between them is testes with a Kolmogorov-Smirnov test - if 
they are \emph{significantly different}, only the top group(s) are reported. 
ie. A more stringent \bold{ks_pval} is more likely to report multiple 
reference groups. 
Note that comparisons with small numbers of 'top' genes will never be 
'significantly' different - these should be taken with a grain of salt 
anyway. This parameter unlikely to need change. Default = 0.01.}

\item{median_rank_threshold}{A threshold for similarity. Number betweeen 0 
(top-rank) and 1.  If the median ranking of the 'top' query genes is 
below this, consider possible similarity. 
Consider reducing for very similar cell clusters. Note that 0.5 is 
theoretically random (but do occur, see detail), so values larger than that
are pointless. Default 0.25.}

\item{de_table.marked}{The output of \code{\link{get_the_up_genes_for_all_possible_groups}} for the contrast of interest.}
}
\value{
A table of automagically-generated labels for each query group, given 
their similarity to reference groups. Short_lab and long_lab are just 
different versions of the name.
}
\description{
Given the comparison between the two dataset, construct some sensible labels 
for the groups in the query group.
}
\details{
This function aims to report a) the top most similar reference group, if 
there's a clear frontrunner, b) A list of multiple similar groups if they 
have similar similarity, or c) 'No similarity', if there is none.

Each group is named according to the following rules:
\enumerate{
  \item There is \emph{one} reference group under the \bold{median_rank_threshold}: Use that one.
  \item There are \emph{multiple} reference group under the
  \bold{median_rank_threshold}: Run a Kolmogorov-Smirnov test on the 
  \bold{rescaled_rank} distribution of query 'top' genes in the 2 most similar reference groups.
  \itemize{
     \item If there is a signifiant (<= \bold{ks_pval}) \emph{difference between 
     the two distributions} - report the first only. Because they're similar,
     but different enough to tell apart.
     \item If the distributions are \emph{not different}, report both. If there are 
     more groups under the \bold{median_rank_threshold} start stepping down 
     group by group to see if there's a difference somewhere (e.g. 3 groups 
     under threshold, but its possible to return only the top 2.).
  }
  \item There is nothing under the \bold{median_rank_threshold} : No simiarity. 
}

The similarity is formatted into a group label. Where there are 
multiple similar groups, they're listed from most to least similar by their 
median ranks.
 
For instance, a query dataset of clusters c1, c2 and c3 againsts a cell-type
labelled reference datatset might get names like:
E.g.
\itemize{
  \item c1:macrophage
  \item c2:endotheial|mesodermal
  \item c3:no_similarity
}
}
\examples{

de_table.demo_query <- contrast_each_group_to_the_rest(demo_query_se, "demo_query")
de_table.demo_ref   <- contrast_each_group_to_the_rest(demo_ref_se,   "demo_ref")
de_table.marked.query_vs_ref <- get_the_up_genes_for_all_possible_groups(de_table.demo_query, de_table.demo_ref, 'demo_query')

make_ref_similarity_names_for_groups_ksfunction(de_table.marked.query_vs_ref,
                              the_test_dataset="demo_query",
                              the_ref_dataset="demo_ref")

}
\seealso{
\code{\link[celaref]{get_the_up_genes_for_all_possible_groups}} To prepare the \bold{de_table.ref.marked} input.
}
