% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{run_pair_test_stats}
\alias{run_pair_test_stats}
\title{run_pair_test_stats}
\usage{
run_pair_test_stats(de_table.ref.marked, the_test_group, groupA, groupB)
}
\arguments{
\item{de_table.ref.marked}{The output of \code{\link{get_the_up_genes_for_all_possible_groups}} for the contrast of interest.}

\item{the_test_group}{Name of the test group in query dataset.}

\item{groupA}{One of the reference group names}

\item{groupB}{Another of the reference group names}
}
\value{
A tibble of wilcox / man-whitneyU test results for this contrast.
}
\description{
Internal function to compare the distribution of a query datasets 'top' 
genes between two different reference datasete groups with a 
Mann–Whitney U test. One directional test if groupA median < group B.
}
\details{
For use by make_ref_similarity_names_for_groups
}
\seealso{
\code{\link[celaref]{make_ref_similarity_names_for_groups}}
}
