% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_labelling_functions.r
\name{get_ranking_and_test_results}
\alias{get_ranking_and_test_results}
\title{get_ranking_and_test_results}
\usage{
get_ranking_and_test_results(de_table.ref.marked, the_test_group,
  the_test_dataset, the_ref_dataset, num_steps, pval = 0.01)
}
\arguments{
\item{de_table.ref.marked}{see 
\link{make_ref_similarity_names_using_marked}}

\item{the_test_group}{The group to calculate the stats on.}

\item{the_test_dataset}{see 
\link{make_ref_similarity_names_using_marked}}

\item{the_ref_dataset}{see 
\link{make_ref_similarity_names_using_marked}}

\item{num_steps}{see 
\link{make_ref_similarity_names_using_marked}}

\item{pval}{see 
\link{make_ref_similarity_names_using_marked}}
}
\value{
Table of similarity contrast results/assigned names 
e.t.c for a single group. 
Used internally for populating mwtest_res_table tables.
}
\description{
Internal function to get reference group similarity contrasts for an 
individual query qroup.
}
\details{
For use by \bold{make_ref_similarity_names_using_marked}, see that function 
for parameter details.
This function just runs this for a single query group \bold{the_test_group}
}
\seealso{
\code{\link{make_ref_similarity_names_using_marked}} 
which calls this.
}
